[SEC-28845] Delete SDKv2 security_monitoring_filter resource, move shared helpers to data source
